:tea: Rewrite command exectutors and add jedis to pom.xml

+++ src/main/java/fr/mrlizzard/wardevil/builder/objects/BuildPlayer.java
@@ -5,6 +5,7 @@ import fr.mrlizzard.wardevil.builder.WardBuilder;
 import fr.mrlizzard.wardevil.builder.uitls.Rank;
 import org.bukkit.entity.Player;
 
+import java.io.File;
 import java.io.FileWriter;
 import java.util.UUID;
 
@@ -48,19 +49,36 @@ public class BuildPlayer {
     }
 
     public void savePlayerConfig() {
+        File playerFile = new File(instance.getDataFolder(), "players/" + uuid.toString() + ".json");
         FileWriter writer;
+        String content;
 
         try {
-            writer = new FileWriter(instance.getDataFolder() + "/players/" + uuid.toString() + ".json");
-
-            try {
-                instance.getGson().toJson(this, writer);
-            } catch (Exception err) {
-                throw new Exception("File " + uuid.toString() + ".json isn't a valid json file.");
-            }
-        } catch (Exception except) {
-            instance.getLog().error(except.getMessage());
+            content = instance.getGson().toJson(this);
+            instance.getLog().warning(content);
+        } catch (Exception exception) {
+            instance.getLog().error(exception.getMessage());
+            return;
         }
+
+        if (!playerFile.exists()) {
+            try {
+                writer = new FileWriter(instance.getDataFolder() + "/players/" + uuid.toString() + ".json");
+
+                try {
+                    writer.write(content);
+                } catch (Exception err) {
+                    throw new Exception("File " + uuid.toString() + ".json isn't a valid json file.");
+                }
+                writer.close();
+            } catch (Exception except) {
+                instance.getLog().error(except.getMessage());
+            }
+
+            return;
+        }
+
+
     }
 
 }
